requirements.txt
Faker
sqlalchemy
psycopg2-binary